Move LazyEntries into training/manifest.py
dataloader.py was accumulating unrelated concerns; the manifest index
is its own thing.
